+++ lib/hooks/auth/index.ts
@@ -1,0 +1,4 @@
+export { useConfirm } from "./useConfirm";
+export { useLogin } from "./useLogin";
+export { useRegister } from "./useRegister";
+export { useResetPassword } from "./useResetPassword";
