+++ lib/hooks/index.ts
@@ -3,4 +3,4 @@ export { useRefresh } from "./useRefresh";
 export { useExternalScript } from "./useExternalScript";
 export { useFormTimer } from "./useFormTimer";
 export { useAccessControl, AccessControlProvider } from "./useAccessControl";
-export { useAuth } from "./useAuth";
+// export { useAuth } from "./auth/useAuth";
